requirements.txt
xmltodict>=0.11
scikit_image
Shapely>=1.6
typing>=3.6.4
easydict>=1.7
tqdm>=4.23.4
matplotlib>=2.2.2
numpy>=1.14
Pillow>=5.4
wget>=3
aggdraw
